--- src/common/components/CoreLayout/Sidebar/types.ts
@@ -1,7 +1,0 @@
-export type SidebarListItem = {
-  label: string;
-  icon: React.ReactNode;
-  href: string;
-  active?: boolean;
-  external?: boolean;
-};
