+++ src/common/components/CoreLayout/index.ts
@@ -1,2 +1,1 @@
 export * from "./CoreLayout";
-export * from "./Sidebar/Sidebar";
