+++ src/common/components/CustomIcon/index.ts
@@ -1,6 +1,7 @@
 export * from "./CustomIcon";
 export * from "./StarLineAltIcon";
 export * from "./AfterclassIcon";
+export * from "./AfterclassText";
 export * from "./BookLineIcon";
 export * from "./DealsIcon";
 export * from "../Logo";
