+++ src/common/components/Logo/Logo.tsx
@@ -1,0 +1,15 @@
+import { AfterclassIcon, AfterclassText } from "../CustomIcon";
+
+export interface LogoProps {
+  hideText?: boolean;
+  hideLogo?: boolean;
+}
+
+export const Logo = ({ hideText = false, hideLogo = false }: LogoProps) => {
+  return (
+    <div className="flex h-6 items-center justify-center gap-x-3">
+      {!hideLogo && <AfterclassIcon size={16} />}
+      {!hideText && <AfterclassText />}
+    </div>
+  );
+};
